main: Move custom palette to styles
